+++ libcorrsk.h
@@ -5,7 +5,7 @@
 
 // just convolution
 template <class T, class U, class V>
-void corrSK(T *X, unsigned Wx, unsigned Hx, U *Krow, U *Kcol, unsigned w, unsigned h, V *Y, unsigned *flop);
+void corrSK(T *X, unsigned Wx, unsigned Hx, U *Krow, U *Kcol, unsigned w, unsigned h, V *Y, unsigned long long *flop);
 
 #include "libcorrsk.cpp"
